feat: V4 Architecture. (#195)
* feat: rewrite page system.

# Conflicts:
#	growingio-autotracker-core/src/main/java/com/growingio/android/sdk/autotrack/inject/ViewChangeProvider.java
#	growingio-autotracker-core/src/main/java/com/growingio/android/sdk/autotrack/inject/ViewClickProvider.java

* 1. 在新设置一个用户信息的场景时（null->userid），由之前的补发VISIT =>不再补发VISIT；
2. 打开收集数据的开关时，强制刷新Session并生成一个VISIT；
3. setLocation 接口不再补发VISIT；
4. 删除 gioid 字段；
5. 删除GlobalSequenceId字段；
6. EventSequenceId 改为全局统计，计(vst、page、click、change、custom)；

* Android SDK Version Upgrade
1. targetSdkVersion -> 33
2. AGP -> 7.4.1
3. can use kotlin but may be have version conflicts;
4. use **libs.version.toml** for version control;
5. android module use namespace replace AndroidManifest.xml;
6. remove gio-sdk:autotracker-cdp and gio-sdk:tracker-cdp;
7. make gio-sdk:tracker deprecated.

* fix warning with disableAutomaticComponentCreation

* rewrite v3.0 ViewNode calculate.

* data transfer use protobuf default.
1. add JsonSerializer for event parser in apt;
2. remove ResourceItemCustomEvent.java;
3. sdk use protobuf default.

* fix event accessibility

* autotracker for v4.0.

* find content in group view

* path get alias first.

* Code Architecture Refactoring.
1. hide Provider is called and delete the singleton of the provider;
2. provider is now in TrackerContext;
3. add shutdown method;
3. user spotless format code;

* fix compiler generate

* fix page path

* page's title will auto get from toolbar.

* move getSupportToolBarTitle to autotrack

* fix eventType judge

* 1. don't replace -
2. publish xIndex to xContent

* fix expandlist path

* page add method:getClassName;
fix:tipview don't show at first time.

* feat:ui test (#197)

* add SDK demo as submodule

* update pre release action

* update runner working-directory

* add ci publishAllToMavenLocal.sh

* update uniqueTag xpath

* update uniqueTag xpath

* fix alertDialog xpath

* fix repairPid invoke before sessionProvider setup

* fix fragment page loss autotrack state.
make activity and fragment weakReference.

* test pass

* apm GMonitor need Application Context

* fix: page tree children duplicate;
fix: checkView2PageElement error;
fix: event's platform always is Android.

* fix: v3 circle select element xpath

* fix: v3 circle select element xpath

* add utilInjector

* spotlessApply

* fix hybrid test

* fix: catch InternalError when starting okhttp thread during runtime shutdown.

* add findViewNodesWithinCircle in ViewNodeRenderer class.

* feat: saas adapter (#199)

* feat: saas adapter

* fix

* fix

* feat: support bridgeWebView

* 1. HybridConfiguration post special version when is downgrade.
2. timezoneOffset is String

* fix startAfterSdkSetup method

* version 4.0.0-SNAPSHOT

* spotlessCheck

* feat: disable collection of androidId by default

* feat: disable requireAppProcesses by default

* 1. hide visitor api
2. update bom version

* fix: V4 circler use Page path replace ui

* fix: getAllWindowViews

* feat:通用属性 (#200)

* feat:setGeneralProps

* check props nullable

* rename remove props api

* fix:remove window from xpath

* rename params projectId->accountId

* make generalProp api called in SDK Main Thread.

* rename module: advert -> ads

* only post index when the value >= 0

* add isRunning method to GrowingTracker

* fix:eventBuilder NPE

* index start with -1

* fix:listview indeedindex

* fix:index start with +1

* upload index when >0

* don't support volley module anymore.

* resolve conflict

* spotless

* fix:http test error

* ci: update to java 11

* feat: abtest (#203)

* feat: abtest module

* flutter 2.0

* fix: when restart webservice reset socket state.

* feat: hybrid bridge add datasourceId

* release 4.0.0

* update android-emulator api level to 33

* update android-emulator

---------

+++ growingio-tracker-core/src/main/java/com/growingio/android/sdk/track/middleware/EventSender.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Beijing Yishu Technology Co., Ltd.
+ * Copyright (C) 2023 Beijing Yishu Technology Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.growingio.android.sdk.track.middleware;
 
 import android.annotation.SuppressLint;
@@ -27,10 +26,10 @@ import android.os.Message;
 
 import androidx.annotation.NonNull;
 
-import com.growingio.android.sdk.TrackerContext;
 import com.growingio.android.sdk.track.ipc.ProcessLock;
 import com.growingio.android.sdk.track.log.Logger;
 import com.growingio.android.sdk.track.modelloader.ModelLoader;
+import com.growingio.android.sdk.track.modelloader.TrackerRegistry;
 import com.growingio.android.sdk.track.utils.NetworkUtil;
 
 import java.text.SimpleDateFormat;
@@ -59,6 +58,8 @@ public class EventSender {
 
     private int mCacheEventNum = 0;
 
+    private final TrackerRegistry mRegistry;
+
     /**
      * 事件发送管理类
      *
@@ -67,8 +68,9 @@ public class EventSender {
      * @param cellularDataLimit  事件发送的移动网络的流量限制，单位 MB
      */
     @SuppressLint("WrongConstant")
-    public EventSender(IEventNetSender sender, long dataUploadInterval, long cellularDataLimit) {
-        mContext = TrackerContext.get().getApplicationContext();
+    public EventSender(Context context, TrackerRegistry registry, IEventNetSender sender, long dataUploadInterval, long cellularDataLimit) {
+        mContext = context.getApplicationContext();
+        mRegistry = registry;
         mCellularDataLimit = cellularDataLimit * 1024L * 1024L;
         mDataUploadInterval = dataUploadInterval * 1000L;
         mEventNetSender = sender;
@@ -80,7 +82,7 @@ public class EventSender {
     }
 
     private ModelLoader<EventDatabase, EventDbResult> getDatabaseModelLoader() {
-        return TrackerContext.get().getRegistry().getModelLoader(EventDatabase.class, EventDbResult.class);
+        return mRegistry.getModelLoader(EventDatabase.class, EventDbResult.class);
     }
 
     private EventDbResult databaseOperation(EventDatabase eventDatabase) {
@@ -220,15 +222,20 @@ public class EventSender {
                 }
                 EventDbResult dbResult = databaseOperation(EventDatabase.query(policy, numOfMaxEventsPerRequest()));
                 if (dbResult.isSuccess() && dbResult.getSum() > 0) {
-                    SendResponse sendResponse = mEventNetSender.send(dbResult.getData(), dbResult.getMediaType());
-                    succeeded = sendResponse.isSucceeded();
-                    if (succeeded) {
-                        String eventType = dbResult.getEventType();
-                        databaseOperation(EventDatabase.delete(dbResult.getLastId(), policy, eventType));
-                        if (networkState.isMobileData()) {
-                            todayBytes(sendResponse.getUsedBytes());
+                    if (mEventNetSender == null) {
+                        succeeded = false;
+                    } else {
+                        SendResponse sendResponse = mEventNetSender.send(dbResult.getData(), dbResult.getMediaType());
+                        succeeded = sendResponse.isSucceeded();
+                        if (succeeded) {
+                            String eventType = dbResult.getEventType();
+                            databaseOperation(EventDatabase.delete(dbResult.getLastId(), policy, eventType));
+                            if (networkState.isMobileData()) {
+                                todayBytes(sendResponse.getUsedBytes());
+                            }
                         }
                     }
+
                 } else {
                     break;
                 }
